Style pages according to contractor designs
In the designs, there is a background image that has to cover
the main top area and whose bottom has to be aligned with the
top of  the gift aid area. Also, the main quote had to be
contained in  the top area, in the limits of the background
image.

I'm not a designer, so after several unsuccessful approaches,
the layout that worked is one in which the form is the main
container, with a page-top and a page-bottom areas.

* The top area has the form elements that go before the gift
aid area, and the quote.

* Tho bottom area has the gift aid area followed by the rest of
the form elements, and the information links

I also kept in mind that the final file size should not be over
around 20k, since there are also custom fonts, images and the
payment-related javascript libraries, that need to be downloaded.

The layout is as follows:

--------------------------------------------
|  form as a container                     |
|  --------------------------------------  |
|  |  page-top (with bg image)          |  |
|  |   ______________   _____________   |  |
|  |  |              |  |            |  |  |
|  |  |  top inputs  |  |   quote    |  |  |
|  |  |______________|  |____________|  |  |
|  |                                    |  |
|  |------------------------------------|  |
|  |  page-bottom                       |  |
|  |   ______________   _____________   |  |
|  |  |              |  |            |  |  |
|  |  |  gift area   |  |   info     |  |  |
|  |  |______________|  |____________|  |  |
|  |  |              |                  |  |
|  |  | other inputs |                  |  |
|  |  |______________|                  |  |
|  |                                    |  |
|  --------------------------------------  |
--------------------------------------------