Add a Heroku client
The Heroku client uploads the new timestamped filenames of the
CSS and JS assets in an upstream-only environment variable.

This allows to synchronize the last filename uploaded to AWS S3
with the path shown in the rendered HTML at Heroku servers.

The client also pushes the local app to Heroku using git push.
There is a way to do this using the API and the gem, but it is
too complicated and with git you can do in in a on-line command.